+++ src/components/Select.jsx
@@ -99,6 +99,7 @@ const ValueContainer = props => {
       {...props}
       innerProps={{
         ...props.innerProps,
+        name: selectProps.name,
         "data-cy": selectProps
           ? `${hyphenize(selectProps.label)}-select-value-container`
           : "select-value-container",
@@ -191,6 +192,7 @@ const Select = ({
         </Label>
       )}
       <Parent
+        blurInputOnSelect={false}
         classNamePrefix="neeto-ui-react-select"
         data-cy={`${hyphenize(label)}-select-container`}
         defaultValue={findInOptions(defaultValue)}